Dockerfile with signature check
It allows to define DEBIAN VERSION (11 or 12),
as well as VERACRYPT_VERSION.

It downloads the corresponding deb, along with its signature.
It checks the veracrypt source FINGERPRINT against a known source, and
it checks the deb signature, before installing it.

If anything goes wrong during the build process, the whole process
fails.